chore: split spammer into multiple files
